routes for getting user playlists and adding song to specific playlist
